+++ src/service/test_support.rs
@@ -32,7 +32,7 @@ impl LLMProvider for NoopLLM {
         _model: &str,
         _messages: &[ChatMessage],
         _tools: &[ToolSchema],
-        _temperature: f32,
+        _temperature: f64,
     ) -> crate::base::Result<LLMResponse> {
         Err(ErrorCode::internal("noop llm"))
     }
@@ -42,7 +42,7 @@ impl LLMProvider for NoopLLM {
         _model: &str,
         _messages: &[ChatMessage],
         _tools: &[ToolSchema],
-        _temperature: f32,
+        _temperature: f64,
     ) -> ResponseStream {
         let (_writer, stream) = ResponseStream::channel(1);
         stream
